refactor(runtime): extract context recovery into agent/turn — #7415
Move the chat_result Err-arm into turn/context_recovery.rs as
record_llm_failure(..) (observer LlmResponse failure + llm_response
failure log) and try_recover_context_overflow(&mut history, ..) -> bool
(fast-trim, then emergency trim, each retrying via the loop; the
unrecoverable case logs and falls through). The loop keeps
`if recovered { continue; } return Err(e);` inline per the RUN_SHEET.
